+++ voxygen/src/settings/gameplay.rs
@@ -14,6 +14,7 @@ pub struct GameplaySettings {
     pub free_look_behavior: PressBehavior,
     pub auto_walk_behavior: PressBehavior,
     pub camera_clamp_behavior: PressBehavior,
+    pub player_physics_behavior: bool,
     pub stop_auto_walk_on_input: bool,
     pub auto_camera: bool,
 }
@@ -30,6 +31,7 @@ impl Default for GameplaySettings {
             free_look_behavior: PressBehavior::Toggle,
             auto_walk_behavior: PressBehavior::Toggle,
             camera_clamp_behavior: PressBehavior::Toggle,
+            player_physics_behavior: false,
             stop_auto_walk_on_input: true,
             auto_camera: false,
         }